+++ android/app-newm/src/main/AndroidManifest.xml
@@ -11,7 +11,6 @@
             android:label="@string/app_name"
             android:roundIcon="@mipmap/ic_launcher_round"
             android:supportsRtl="true"
-           android:networkSecurityConfig="@xml/network_security_config"
             android:theme="@style/Theme.ProjectNewm.SplashScreen">
 
         <activity
