+++ android/core-resources/src/debug/res/drawable/ic_launcher_background.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <size
+        android:width="108dp"
+        android:height="108dp" />
+    <solid android:color="#fff" />
+</shape>
